Fix ATmega328P clock which was not consistent with config of 1000 tick/s
ATmega328P Timer2 has no automatic reload so we need to manually
reload it in the ISR. The new timer settings work with a main clock
of 16 MHz which is the default clock on some dev. boards like the
Arduino UNO. For a lower main clock, the user might need to disable
some options for keeping nOS_Tick() from bloating the whole system.